Add multi unit and invisibility support
Mutli unit support will now allow the script to utilize all phase boots
available to the local player. This might be a problem if another
player shared control of their units, but it is rather a rare case.

Invisibility support will now force script to do nothing if the phase
boots owner is invisible. But due to delays with the server there might
be rare occurances of phase breaking the invisibility right at the
beginning of the invisibility. This could be avoided with fade timed
invisibility sources but it is unavoidable with instant invisibilities.
It will probably contain this bug until there is an easy way to get
when the unit will go invisible due to the fade time, but before that
I feel it is an overly complicated thing to add to this script.